+++ src/backend/dx11/src/lib.rs
@@ -21,31 +21,40 @@ extern crate bitflags;
 #[macro_use]
 extern crate log;
 
+use crate::{debug::set_debug_name, device::DepthStencilState};
 use auxil::ShaderStage;
 use hal::{
     adapter, buffer, command, format, image, memory, pass, pso, query, queue, window, DrawCount,
     IndexCount, InstanceCount, Limits, TaskCount, VertexCount, VertexOffset, WorkGroupCount,
 };
 use range_alloc::RangeAllocator;
-use crate::{
-    device::DepthStencilState,
-    debug::set_debug_name,
-};
 
-use winapi::{shared::{
-    dxgi::{IDXGIAdapter, IDXGIFactory, IDXGISwapChain},
-    dxgiformat,
-    minwindef::{FALSE, HMODULE, UINT},
-    windef::{HWND, RECT},
-    winerror,
-}, um::{d3d11, d3d11_1, d3dcommon, winuser::GetClientRect}, Interface as _};
+use winapi::{
+    shared::{
+        dxgi::{IDXGIAdapter, IDXGIFactory, IDXGISwapChain},
+        dxgiformat,
+        minwindef::{FALSE, HMODULE, UINT},
+        windef::{HWND, RECT},
+        winerror,
+    },
+    um::{d3d11, d3d11_1, d3dcommon, winuser::GetClientRect},
+    Interface as _,
+};
 
 use wio::com::ComPtr;
 
 use arrayvec::ArrayVec;
 use parking_lot::{Condvar, Mutex, RwLock};
 
-use std::{borrow::Borrow, cell::RefCell, fmt, mem, ops::Range, os::raw::c_void, ptr, sync::{Arc, Weak}};
+use std::{
+    borrow::Borrow,
+    cell::RefCell,
+    fmt, mem,
+    ops::Range,
+    os::raw::c_void,
+    ptr,
+    sync::{Arc, Weak},
+};
 
 macro_rules! debug_scope {
     ($context:expr, $($arg:tt)+) => ({
@@ -149,28 +158,28 @@ fn get_features(
 
     features.set(
         hal::Features::TEXTURE_DESCRIPTOR_ARRAY
-        | hal::Features::FULL_DRAW_INDEX_U32
-        | hal::Features::GEOMETRY_SHADER,
-        feature_level >= d3dcommon::D3D_FEATURE_LEVEL_10_0
+            | hal::Features::FULL_DRAW_INDEX_U32
+            | hal::Features::GEOMETRY_SHADER,
+        feature_level >= d3dcommon::D3D_FEATURE_LEVEL_10_0,
     );
 
     features.set(
         hal::Features::IMAGE_CUBE_ARRAY,
-        feature_level >= d3dcommon::D3D_FEATURE_LEVEL_10_1
+        feature_level >= d3dcommon::D3D_FEATURE_LEVEL_10_1,
     );
 
     features.set(
         hal::Features::VERTEX_STORES_AND_ATOMICS
-        | hal::Features::FRAGMENT_STORES_AND_ATOMICS
-        | hal::Features::FORMAT_BC
-        | hal::Features::TESSELLATION_SHADER
-        | hal::Features::DRAW_INDIRECT_FIRST_INSTANCE,
-        feature_level >= d3dcommon::D3D_FEATURE_LEVEL_11_0
+            | hal::Features::FRAGMENT_STORES_AND_ATOMICS
+            | hal::Features::FORMAT_BC
+            | hal::Features::TESSELLATION_SHADER
+            | hal::Features::DRAW_INDIRECT_FIRST_INSTANCE,
+        feature_level >= d3dcommon::D3D_FEATURE_LEVEL_11_0,
     );
 
     features.set(
         hal::Features::LOGIC_OP, // TODO: Optional at 10_0 -> 11_0
-        feature_level >= d3dcommon::D3D_FEATURE_LEVEL_11_1
+        feature_level >= d3dcommon::D3D_FEATURE_LEVEL_11_1,
     );
 
     features
@@ -198,8 +207,7 @@ fn get_limits(feature_level: d3dcommon::D3D_FEATURE_LEVEL) -> hal::Limits {
     };
 
     let max_texture_cube_dimension = match feature_level {
-        d3dcommon::D3D_FEATURE_LEVEL_9_1
-        | d3dcommon::D3D_FEATURE_LEVEL_9_2 => 512,
+        d3dcommon::D3D_FEATURE_LEVEL_9_1 | d3dcommon::D3D_FEATURE_LEVEL_9_2 => 512,
         _ => max_texture_uv_dimension,
     };
 
@@ -236,9 +244,7 @@ fn get_limits(feature_level: d3dcommon::D3D_FEATURE_LEVEL) -> hal::Limits {
         | d3dcommon::D3D_FEATURE_LEVEL_9_3
         | d3dcommon::D3D_FEATURE_LEVEL_10_0 => 0b0001, // Conservative, MSAA isn't required.
         d3dcommon::D3D_FEATURE_LEVEL_10_1 => 0b0101, // Optimistic, 4xMSAA is required on all formats _but_ RGBA32.
-        d3dcommon::D3D_FEATURE_LEVEL_11_0
-        | d3dcommon::D3D_FEATURE_LEVEL_11_1
-        | _ => 0b1101, // Optimistic, 8xMSAA and 4xMSAA is required on all formats _but_ RGBA32 which requires 4x.
+        d3dcommon::D3D_FEATURE_LEVEL_11_0 | d3dcommon::D3D_FEATURE_LEVEL_11_1 | _ => 0b1101, // Optimistic, 8xMSAA and 4xMSAA is required on all formats _but_ RGBA32 which requires 4x.
     };
 
     let max_constant_buffers = d3d11::D3D11_COMMONSHADER_CONSTANT_BUFFER_API_SLOT_COUNT - 1;
@@ -254,7 +260,8 @@ fn get_limits(feature_level: d3dcommon::D3D_FEATURE_LEVEL) -> hal::Limits {
         max_per_stage_descriptor_uniform_buffers: max_constant_buffers as _,
         max_per_stage_descriptor_storage_buffers: max_buffer_uav,
         max_per_stage_descriptor_storage_images: max_image_uav,
-        max_per_stage_descriptor_sampled_images: d3d11::D3D11_COMMONSHADER_INPUT_RESOURCE_REGISTER_COUNT as _,
+        max_per_stage_descriptor_sampled_images:
+            d3d11::D3D11_COMMONSHADER_INPUT_RESOURCE_REGISTER_COUNT as _,
         max_descriptor_set_uniform_buffers_dynamic: max_constant_buffers as _,
         max_descriptor_set_storage_buffers_dynamic: 0, // TODO: Implement dynamic offsets for storage buffers
         max_texel_elements: max_texture_uv_dimension as _, //TODO
@@ -280,10 +287,11 @@ fn get_limits(feature_level: d3dcommon::D3D_FEATURE_LEVEL) -> hal::Limits {
         ], // TODO
         max_vertex_input_attribute_offset: 255, // TODO
         max_vertex_input_attributes: max_input_slots,
-        max_vertex_input_binding_stride: d3d11::D3D11_REQ_MULTI_ELEMENT_STRUCTURE_SIZE_IN_BYTES as _,
+        max_vertex_input_binding_stride: d3d11::D3D11_REQ_MULTI_ELEMENT_STRUCTURE_SIZE_IN_BYTES
+            as _,
         max_vertex_input_bindings: d3d11::D3D11_IA_VERTEX_INPUT_RESOURCE_SLOT_COUNT as _, // TODO: verify same as attributes
-        max_vertex_output_components: d3d11::D3D11_VS_OUTPUT_REGISTER_COUNT as _, // TODO
-        min_texel_buffer_offset_alignment: 1,                                     // TODO
+        max_vertex_output_components: d3d11::D3D11_VS_OUTPUT_REGISTER_COUNT as _,         // TODO
+        min_texel_buffer_offset_alignment: 1,                                             // TODO
         min_uniform_buffer_offset_alignment: 16,
         min_storage_buffer_offset_alignment: 16, // TODO
         framebuffer_color_sample_counts: max_samples,
@@ -674,7 +682,12 @@ impl adapter::PhysicalDevice<Backend> for PhysicalDevice {
                 return Err(hal::device::CreationError::InitializationFailed);
             }
 
-            info!("feature level={:x}=FL{}_{}", feature_level, feature_level >> 12, feature_level >> 8 & 0xF);
+            info!(
+                "feature level={:x}=FL{}_{}",
+                feature_level,
+                feature_level >> 12,
+                feature_level >> 8 & 0xF
+            );
 
             (ComPtr::from_raw(device), ComPtr::from_raw(cxt))
         };
@@ -878,8 +891,7 @@ impl window::Surface<Backend> for Surface {
         // NOTE: some swap effects affect msaa capabilities..
         // TODO: _DISCARD swap effects can only have one image?
         window::SurfaceCapabilities {
-            present_modes: window::PresentMode::IMMEDIATE
-                | window::PresentMode::FIFO,
+            present_modes: window::PresentMode::IMMEDIATE | window::PresentMode::FIFO,
             composite_alpha_modes: window::CompositeAlphaMode::OPAQUE, //TODO
             image_count: 1..=16,                                       // TODO:
             current_extent,
@@ -1157,25 +1169,17 @@ impl RenderPassCache {
         let color_views = subpass
             .color_attachments
             .iter()
-            .map(|&(id, _)| {
-                self.framebuffer.attachments[id]
-                    .rtv_handle
-                    .unwrap()
-            })
+            .map(|&(id, _)| self.framebuffer.attachments[id].rtv_handle.unwrap())
             .collect::<Vec<_>>();
         let (ds_view, rods_view) = match subpass.depth_stencil_attachment {
             Some((id, _)) => {
                 let attachment = &self.framebuffer.attachments[id];
-                let ds_view = attachment
-                    .dsv_handle
-                    .unwrap();
+                let ds_view = attachment.dsv_handle.unwrap();
 
-                let rods_view = attachment
-                    .rodsv_handle
-                    .unwrap();
+                let rods_view = attachment.rodsv_handle.unwrap();
 
                 (Some(ds_view), Some(rods_view))
-            },
+            }
             None => (None, None),
         };
 
@@ -1186,7 +1190,11 @@ impl RenderPassCache {
     fn resolve_msaa(&mut self, context: &ComPtr<d3d11::ID3D11DeviceContext>) {
         let subpass: &SubpassDesc = &self.render_pass.subpasses[self.current_subpass as usize];
 
-        for (&(color_id, _), &(resolve_id, _)) in subpass.color_attachments.iter().zip(subpass.resolve_attachments.iter()) {
+        for (&(color_id, _), &(resolve_id, _)) in subpass
+            .color_attachments
+            .iter()
+            .zip(subpass.resolve_attachments.iter())
+        {
             if color_id == pass::ATTACHMENT_UNUSED || resolve_id == pass::ATTACHMENT_UNUSED {
                 continue;
             }
@@ -1198,15 +1206,21 @@ impl RenderPassCache {
             let mut resolve_resource: *mut d3d11::ID3D11Resource = ptr::null_mut();
 
             unsafe {
-                (&*color_framebuffer.rtv_handle.expect("Framebuffer must have COLOR_ATTACHMENT usage")).GetResource(&mut color_resource as *mut *mut _);
-                (&*resolve_framebuffer.rtv_handle.expect("Resolve texture must have COLOR_ATTACHMENT usage")).GetResource(&mut resolve_resource as *mut *mut _);
+                (&*color_framebuffer
+                    .rtv_handle
+                    .expect("Framebuffer must have COLOR_ATTACHMENT usage"))
+                    .GetResource(&mut color_resource as *mut *mut _);
+                (&*resolve_framebuffer
+                    .rtv_handle
+                    .expect("Resolve texture must have COLOR_ATTACHMENT usage"))
+                    .GetResource(&mut resolve_resource as *mut *mut _);
 
                 context.ResolveSubresource(
                     resolve_resource,
                     resolve_framebuffer.subresource,
                     color_resource,
                     color_framebuffer.subresource,
-                    conv::map_format(color_framebuffer.format).unwrap()
+                    conv::map_format(color_framebuffer.format).unwrap(),
                 );
 
                 (&*color_resource).Release();
@@ -1400,7 +1414,7 @@ impl CommandBufferState {
         &mut self,
         render_targets: &[*mut d3d11::ID3D11RenderTargetView],
         depth_target: Option<*mut d3d11::ID3D11DepthStencilView>,
-        readonly_depth_target: Option<*mut d3d11::ID3D11DepthStencilView>
+        readonly_depth_target: Option<*mut d3d11::ID3D11DepthStencilView>,
     ) {
         for (idx, &rt) in render_targets.iter().enumerate() {
             self.render_targets[idx] = rt;
@@ -1410,11 +1424,15 @@ impl CommandBufferState {
         self.depth_target = depth_target;
         self.readonly_depth_target = readonly_depth_target;
 
-        self.dirty_flag.insert(DirtyStateFlag::RENDER_TARGETS_AND_UAVS);
+        self.dirty_flag
+            .insert(DirtyStateFlag::RENDER_TARGETS_AND_UAVS);
     }
 
     pub fn bind_render_targets(&mut self, context: &ComPtr<d3d11::ID3D11DeviceContext>) {
-        if !self.dirty_flag.contains(DirtyStateFlag::RENDER_TARGETS_AND_UAVS) {
+        if !self
+            .dirty_flag
+            .contains(DirtyStateFlag::RENDER_TARGETS_AND_UAVS)
+        {
             return;
         }
 
@@ -1422,7 +1440,8 @@ impl CommandBufferState {
             self.readonly_depth_target
         } else {
             self.depth_target
-        }.unwrap_or(ptr::null_mut());
+        }
+        .unwrap_or(ptr::null_mut());
 
         let uav_start_index = d3d11::D3D11_PS_CS_UAV_REGISTER_COUNT - self.uav_len;
 
@@ -1446,7 +1465,8 @@ impl CommandBufferState {
             };
         }
 
-        self.dirty_flag.remove(DirtyStateFlag::RENDER_TARGETS_AND_UAVS);
+        self.dirty_flag
+            .remove(DirtyStateFlag::RENDER_TARGETS_AND_UAVS);
     }
 
     pub fn set_blend_factor(&mut self, factor: [f32; 4]) {
@@ -1482,7 +1502,10 @@ impl CommandBufferState {
     }
 
     pub fn bind_depth_stencil_state(&mut self, context: &ComPtr<d3d11::ID3D11DeviceContext>) {
-        if !self.dirty_flag.contains(DirtyStateFlag::DEPTH_STENCIL_STATE) {
+        if !self
+            .dirty_flag
+            .contains(DirtyStateFlag::DEPTH_STENCIL_STATE)
+        {
             return;
         }
 
@@ -1530,18 +1553,19 @@ impl CommandBufferState {
         }
 
         // If we don't have depth stencil state, we use the old value, so we don't bother changing anything.
-        let depth_target_read_only =
-            pipeline
-                .depth_stencil_state
-                .as_ref()
-                .map_or(self.depth_target_read_only, |ds| ds.read_only);
+        let depth_target_read_only = pipeline
+            .depth_stencil_state
+            .as_ref()
+            .map_or(self.depth_target_read_only, |ds| ds.read_only);
 
         if self.depth_target_read_only != depth_target_read_only {
             self.depth_target_read_only = depth_target_read_only;
-            self.dirty_flag.insert(DirtyStateFlag::RENDER_TARGETS_AND_UAVS);
+            self.dirty_flag
+                .insert(DirtyStateFlag::RENDER_TARGETS_AND_UAVS);
         }
 
-        self.dirty_flag.insert(DirtyStateFlag::GRAPHICS_PIPELINE | DirtyStateFlag::DEPTH_STENCIL_STATE);
+        self.dirty_flag
+            .insert(DirtyStateFlag::GRAPHICS_PIPELINE | DirtyStateFlag::DEPTH_STENCIL_STATE);
 
         self.graphics_pipeline = Some(pipeline);
     }
@@ -1569,28 +1593,40 @@ impl CommandBufferState {
                 context.VSSetShader(pipeline.vs.as_raw(), ptr::null_mut(), 0);
 
                 if self.dirty_flag.contains(DirtyStateFlag::PIPELINE_PS) {
-                    let ps = pipeline.ps.as_ref().map_or(ptr::null_mut(), |ps| ps.as_raw());
+                    let ps = pipeline
+                        .ps
+                        .as_ref()
+                        .map_or(ptr::null_mut(), |ps| ps.as_raw());
                     context.PSSetShader(ps, ptr::null_mut(), 0);
 
                     self.dirty_flag.remove(DirtyStateFlag::PIPELINE_PS)
                 }
 
                 if self.dirty_flag.contains(DirtyStateFlag::PIPELINE_GS) {
-                    let gs = pipeline.gs.as_ref().map_or(ptr::null_mut(), |gs| gs.as_raw());
+                    let gs = pipeline
+                        .gs
+                        .as_ref()
+                        .map_or(ptr::null_mut(), |gs| gs.as_raw());
                     context.GSSetShader(gs, ptr::null_mut(), 0);
 
                     self.dirty_flag.remove(DirtyStateFlag::PIPELINE_GS)
                 }
 
                 if self.dirty_flag.contains(DirtyStateFlag::PIPELINE_HS) {
-                    let hs = pipeline.hs.as_ref().map_or(ptr::null_mut(), |hs| hs.as_raw());
+                    let hs = pipeline
+                        .hs
+                        .as_ref()
+                        .map_or(ptr::null_mut(), |hs| hs.as_raw());
                     context.HSSetShader(hs, ptr::null_mut(), 0);
 
                     self.dirty_flag.remove(DirtyStateFlag::PIPELINE_HS)
                 }
 
                 if self.dirty_flag.contains(DirtyStateFlag::PIPELINE_DS) {
-                    let ds = pipeline.ds.as_ref().map_or(ptr::null_mut(), |ds| ds.as_raw());
+                    let ds = pipeline
+                        .ds
+                        .as_ref()
+                        .map_or(ptr::null_mut(), |ds| ds.as_raw());
                     context.DSSetShader(ds, ptr::null_mut(), 0);
 
                     self.dirty_flag.remove(DirtyStateFlag::PIPELINE_DS)
@@ -1622,7 +1658,8 @@ impl CommandBufferState {
     }
 }
 
-type PerConstantBufferVec<T> = ArrayVec<[T; d3d11::D3D11_COMMONSHADER_CONSTANT_BUFFER_API_SLOT_COUNT as _]>;
+type PerConstantBufferVec<T> =
+    ArrayVec<[T; d3d11::D3D11_COMMONSHADER_CONSTANT_BUFFER_API_SLOT_COUNT as _]>;
 
 fn generate_graphics_dynamic_constant_buffer_offsets<'a>(
     bindings: impl IntoIterator<Item = &'a pso::DescriptorSetLayoutBinding>,
@@ -1637,9 +1674,10 @@ fn generate_graphics_dynamic_constant_buffer_offsets<'a>(
     for binding in bindings.into_iter() {
         match binding.ty {
             pso::DescriptorType::Buffer {
-                format: pso::BufferDescriptorFormat::Structured {
-                    dynamic_offset: true
-                },
+                format:
+                    pso::BufferDescriptorFormat::Structured {
+                        dynamic_offset: true,
+                    },
                 ty: pso::BufferDescriptorType::Uniform,
             } => {
                 let offset = offset_iter.next().unwrap();
@@ -1648,30 +1686,38 @@ fn generate_graphics_dynamic_constant_buffer_offsets<'a>(
                     vs_offsets.push(offset / 16)
                 };
 
-                if binding.stage_flags.contains(pso::ShaderStageFlags::FRAGMENT) {
+                if binding
+                    .stage_flags
+                    .contains(pso::ShaderStageFlags::FRAGMENT)
+                {
                     fs_offsets.push(offset / 16)
                 };
                 exists_dynamic_constant_buffer = true;
             }
             pso::DescriptorType::Buffer {
-                format: pso::BufferDescriptorFormat::Structured {
-                    dynamic_offset: false
-                },
+                format:
+                    pso::BufferDescriptorFormat::Structured {
+                        dynamic_offset: false,
+                    },
                 ty: pso::BufferDescriptorType::Uniform,
             } => {
                 if binding.stage_flags.contains(pso::ShaderStageFlags::VERTEX) {
                     vs_offsets.push(0)
                 };
 
-                if binding.stage_flags.contains(pso::ShaderStageFlags::FRAGMENT) {
+                if binding
+                    .stage_flags
+                    .contains(pso::ShaderStageFlags::FRAGMENT)
+                {
                     fs_offsets.push(0)
                 };
             }
             pso::DescriptorType::Buffer {
                 ty: pso::BufferDescriptorType::Storage { .. },
-                format: pso::BufferDescriptorFormat::Structured {
-                    dynamic_offset: true
-                },
+                format:
+                    pso::BufferDescriptorFormat::Structured {
+                        dynamic_offset: true,
+                    },
             } => {
                 // TODO: Storage buffer offsets require new buffer views with correct sizes.
                 //       Might also require D3D11_BUFFEREX_SRV to act like RBA is happening.
@@ -1701,9 +1747,10 @@ fn generate_compute_dynamic_constant_buffer_offsets<'a>(
     for binding in bindings.into_iter() {
         match binding.ty {
             pso::DescriptorType::Buffer {
-                format: pso::BufferDescriptorFormat::Structured {
-                    dynamic_offset: true
-                },
+                format:
+                    pso::BufferDescriptorFormat::Structured {
+                        dynamic_offset: true,
+                    },
                 ty: pso::BufferDescriptorType::Uniform,
             } => {
                 let offset = offset_iter.next().unwrap();
@@ -1715,9 +1762,10 @@ fn generate_compute_dynamic_constant_buffer_offsets<'a>(
                 exists_dynamic_constant_buffer = true;
             }
             pso::DescriptorType::Buffer {
-                format: pso::BufferDescriptorFormat::Structured {
-                    dynamic_offset: false
-                },
+                format:
+                    pso::BufferDescriptorFormat::Structured {
+                        dynamic_offset: false,
+                    },
                 ty: pso::BufferDescriptorType::Uniform,
             } => {
                 if binding.stage_flags.contains(pso::ShaderStageFlags::COMPUTE) {
@@ -1726,9 +1774,10 @@ fn generate_compute_dynamic_constant_buffer_offsets<'a>(
             }
             pso::DescriptorType::Buffer {
                 ty: pso::BufferDescriptorType::Storage { .. },
-                format: pso::BufferDescriptorFormat::Structured {
-                    dynamic_offset: true
-                },
+                format:
+                    pso::BufferDescriptorFormat::Structured {
+                        dynamic_offset: true,
+                    },
             } => {
                 // TODO: Storage buffer offsets require new buffer views with correct sizes.
                 //       Might also require D3D11_BUFFEREX_SRV to act like RBA is happening.
@@ -1790,8 +1839,7 @@ impl CommandBuffer {
     ) -> Self {
         let (context, context1) = if let Some(device1) = device1 {
             let mut context1: *mut d3d11_1::ID3D11DeviceContext1 = ptr::null_mut();
-            let hr =
-                unsafe { device1.CreateDeferredContext1(0, &mut context1 as *mut *mut _) };
+            let hr = unsafe { device1.CreateDeferredContext1(0, &mut context1 as *mut *mut _) };
             assert_eq!(hr, winerror::S_OK);
 
             let context1 = unsafe { ComPtr::from_raw(context1) };
@@ -1800,8 +1848,7 @@ impl CommandBuffer {
             (context, Some(context1))
         } else {
             let mut context: *mut d3d11::ID3D11DeviceContext = ptr::null_mut();
-            let hr =
-                unsafe { device.CreateDeferredContext(0, &mut context as *mut *mut _) };
+            let hr = unsafe { device.CreateDeferredContext(0, &mut context as *mut *mut _) };
             assert_eq!(hr, winerror::S_OK);
 
             let context = unsafe { ComPtr::from_raw(context) };
@@ -1910,17 +1957,17 @@ impl command::CommandBuffer<Backend> for CommandBuffer {
         self.context.VSSetConstantBuffers(
             d3d11::D3D11_COMMONSHADER_CONSTANT_BUFFER_API_SLOT_COUNT - 1,
             1,
-            &raw_push_constant_buffer as *const _
+            &raw_push_constant_buffer as *const _,
         );
         self.context.PSSetConstantBuffers(
             d3d11::D3D11_COMMONSHADER_CONSTANT_BUFFER_API_SLOT_COUNT - 1,
             1,
-            &raw_push_constant_buffer as *const _
+            &raw_push_constant_buffer as *const _,
         );
         self.context.CSSetConstantBuffers(
             d3d11::D3D11_COMMONSHADER_CONSTANT_BUFFER_API_SLOT_COUNT - 1,
             1,
-            &raw_push_constant_buffer as *const _
+            &raw_push_constant_buffer as *const _,
         );
     }
 
@@ -2351,7 +2398,7 @@ impl command::CommandBuffer<Backend> for CommandBuffer {
             let (vs_offsets, fs_offsets) = generate_graphics_dynamic_constant_buffer_offsets(
                 &*set.layout.bindings,
                 &mut offset_iter,
-                self.context1.is_some()
+                self.context1.is_some(),
             );
 
             if let Some(rd) = info.registers.vs.c.as_some() {
@@ -2365,7 +2412,11 @@ impl command::CommandBuffer<Backend> for CommandBuffer {
                         let null_cbuf = [ptr::null_mut::<d3d11::ID3D11Buffer>()];
                         context1.VSSetConstantBuffers(start_slot, 1, &null_cbuf as *const _);
                         if num_buffers > 1 {
-                            context1.VSSetConstantBuffers(start_slot + num_buffers - 1, 1, &null_cbuf as *const _);
+                            context1.VSSetConstantBuffers(
+                                start_slot + num_buffers - 1,
+                                1,
+                                &null_cbuf as *const _,
+                            );
                         }
                     }
 
@@ -2382,7 +2433,7 @@ impl command::CommandBuffer<Backend> for CommandBuffer {
                     self.context.VSSetConstantBuffers(
                         start_slot,
                         num_buffers,
-                        constant_buffers as *const *mut _
+                        constant_buffers as *const *mut _,
                     );
                 }
             }
@@ -2412,7 +2463,11 @@ impl command::CommandBuffer<Backend> for CommandBuffer {
                         let null_cbuf = [ptr::null_mut::<d3d11::ID3D11Buffer>()];
                         context1.PSSetConstantBuffers(start_slot, 1, &null_cbuf as *const _);
                         if num_buffers > 1 {
-                            context1.PSSetConstantBuffers(start_slot + num_buffers - 1, 1, &null_cbuf as *const _);
+                            context1.PSSetConstantBuffers(
+                                start_slot + num_buffers - 1,
+                                1,
+                                &null_cbuf as *const _,
+                            );
                         }
                     }
 
@@ -2427,7 +2482,7 @@ impl command::CommandBuffer<Backend> for CommandBuffer {
                     self.context.PSSetConstantBuffers(
                         start_slot,
                         num_buffers,
-                        constant_buffers as *const *mut _
+                        constant_buffers as *const *mut _,
                     );
                 }
             }
@@ -2448,15 +2503,20 @@ impl command::CommandBuffer<Backend> for CommandBuffer {
 
             // UAVs going to the graphics pipeline are always treated as pixel shader bindings.
             if let Some(rd) = info.registers.ps.u.as_some() {
-            	// We bind UAVs in inverse order from the top to prevent invalidation
-            	// when the render target count changes.
+                // We bind UAVs in inverse order from the top to prevent invalidation
+                // when the render target count changes.
                 for idx in (0..(rd.count)).rev() {
                     let ptr = (*set.handles.offset(rd.pool_offset as isize + idx as isize)).0;
-                    let uav_register = d3d11::D3D11_PS_CS_UAV_REGISTER_COUNT - 1 - rd.res_index as u32 - idx as u32;
+                    let uav_register = d3d11::D3D11_PS_CS_UAV_REGISTER_COUNT
+                        - 1
+                        - rd.res_index as u32
+                        - idx as u32;
                     self.cache.uavs[uav_register as usize] = ptr as *mut _;
                 }
                 self.cache.uav_len = (rd.res_index + rd.count) as u32;
-                self.cache.dirty_flag.insert(DirtyStateFlag::RENDER_TARGETS_AND_UAVS);
+                self.cache
+                    .dirty_flag
+                    .insert(DirtyStateFlag::RENDER_TARGETS_AND_UAVS);
             }
         }
 
@@ -2532,7 +2592,7 @@ impl command::CommandBuffer<Backend> for CommandBuffer {
             let cs_offsets = generate_compute_dynamic_constant_buffer_offsets(
                 &*set.layout.bindings,
                 &mut offset_iter,
-                self.context1.is_some()
+                self.context1.is_some(),
             );
 
             if let Some(rd) = info.registers.cs.c.as_some() {
@@ -2546,7 +2606,11 @@ impl command::CommandBuffer<Backend> for CommandBuffer {
                         let null_cbuf = [ptr::null_mut::<d3d11::ID3D11Buffer>()];
                         context1.CSSetConstantBuffers(start_slot, 1, &null_cbuf as *const _);
                         if num_buffers > 1 {
-                            context1.CSSetConstantBuffers(start_slot + num_buffers - 1, 1, &null_cbuf as *const _);
+                            context1.CSSetConstantBuffers(
+                                start_slot + num_buffers - 1,
+                                1,
+                                &null_cbuf as *const _,
+                            );
                         }
                     }
 
@@ -2563,7 +2627,7 @@ impl command::CommandBuffer<Backend> for CommandBuffer {
                     self.context.CSSetConstantBuffers(
                         start_slot,
                         num_buffers,
-                        constant_buffers as *const *mut _
+                        constant_buffers as *const *mut _,
                     );
                 }
             }
@@ -2737,10 +2801,8 @@ impl command::CommandBuffer<Backend> for CommandBuffer {
         _stride: u32,
     ) {
         assert_eq!(draw_count, 1, "DX11 doesn't support MULTI_DRAW_INDIRECT");
-        self.context.DrawInstancedIndirect(
-            buffer.internal.raw,
-            offset as _,
-        );
+        self.context
+            .DrawInstancedIndirect(buffer.internal.raw, offset as _);
     }
 
     unsafe fn draw_indexed_indirect(
@@ -2751,10 +2813,8 @@ impl command::CommandBuffer<Backend> for CommandBuffer {
         _stride: u32,
     ) {
         assert_eq!(draw_count, 1, "DX11 doesn't support MULTI_DRAW_INDIRECT");
-        self.context.DrawIndexedInstancedIndirect(
-            buffer.internal.raw,
-            offset as _,
-        );
+        self.context
+            .DrawIndexedInstancedIndirect(buffer.internal.raw, offset as _);
     }
 
     unsafe fn draw_indirect_count(
@@ -2989,10 +3049,13 @@ impl MemoryInvalidate {
         context: &ComPtr<d3d11::ID3D11DeviceContext>,
         buffer: *mut d3d11::ID3D11Buffer,
         host_range: Range<u64>,
-        buffer_range: Range<u64>
+        buffer_range: Range<u64>,
     ) {
         // Range<u64> doesn't impl `len` for some bizzare reason relating to underflow
-        debug_assert_eq!(host_range.end - host_range.start, buffer_range.end - buffer_range.start);
+        debug_assert_eq!(
+            host_range.end - host_range.start,
+            buffer_range.end - buffer_range.start
+        );
 
         unsafe {
             context.CopySubresourceRegion(
@@ -3210,7 +3273,11 @@ impl hal::pool::CommandPool<Backend> for CommandPool {
     }
 
     unsafe fn allocate_one(&mut self, _level: command::Level) -> CommandBuffer {
-        CommandBuffer::create_deferred(&self.device, self.device1.as_deref(), Arc::clone(&self.internal))
+        CommandBuffer::create_deferred(
+            &self.device,
+            self.device1.as_deref(),
+            Arc::clone(&self.internal),
+        )
     }
 
     unsafe fn free<I>(&mut self, _cbufs: I)
@@ -3352,7 +3419,7 @@ pub struct InternalImage {
     /// Contains RTVs for all subresources
     render_target_views: Vec<ComPtr<d3d11::ID3D11RenderTargetView>>,
 
-    debug_name: Option<String>
+    debug_name: Option<String>,
 }
 
 impl InternalImage {
@@ -3432,7 +3499,7 @@ impl Clone for ImageView {
             dsv_handle: self.dsv_handle.clone(),
             rodsv_handle: self.rodsv_handle.clone(),
             uav_handle: self.uav_handle.clone(),
-            owned: false
+            owned: false,
         }
     }
 }
@@ -3610,7 +3677,12 @@ impl<T> MultiStageData<RegisterData<T>> {
 }
 
 impl MultiStageData<RegisterData<DescriptorIndex>> {
-    fn add_content_many(&mut self, content: DescriptorContent, stages: pso::ShaderStageFlags, count: DescriptorIndex) {
+    fn add_content_many(
+        &mut self,
+        content: DescriptorContent,
+        stages: pso::ShaderStageFlags,
+        count: DescriptorIndex,
+    ) {
         if stages.contains(pso::ShaderStageFlags::VERTEX) {
             self.vs.add_content_many(content, count);
         }