+++ src/backend/metal/src/command.rs
@@ -10,7 +10,7 @@ use hal::{
     buffer, command as com,
     device::OutOfMemory,
     format::{Aspects, FormatDesc},
-    image::{Extent, Filter, Layout, Level, SubresourceRange},
+    image::{Extent, Filter, Layout, Level, NumSamples, SubresourceRange},
     memory,
     pass::AttachmentLoadOp,
     pso, query,
@@ -250,6 +250,7 @@ struct SubpassInfo {
     descriptor: metal::RenderPassDescriptor,
     combined_aspects: Aspects,
     formats: native::SubpassFormats,
+    sample_count: NumSamples,
 }
 
 #[derive(Debug, Default)]
@@ -300,6 +301,7 @@ struct State {
     target_aspects: Aspects,
     target_extent: Extent,
     target_formats: native::SubpassFormats,
+    target_samples: NumSamples,
     visibility_query: (metal::MTLVisibilityResultMode, buffer::Offset),
     pending_subpasses: Vec<SubpassInfo>,
     descriptor_sets: ArrayVec<[DescriptorSetInfo; MAX_BOUND_DESCRIPTOR_SETS]>,
@@ -2358,6 +2360,7 @@ impl hal::pool::CommandPool<Backend> for CommandPool {
                 target_aspects: Aspects::empty(),
                 target_extent: Extent::default(),
                 target_formats: native::SubpassFormats::default(),
+                target_samples: 0,
                 visibility_query: (metal::MTLVisibilityResultMode::Disabled, 0),
                 pending_subpasses: Vec::new(),
                 descriptor_sets: (0..MAX_BOUND_DESCRIPTOR_SETS)
@@ -2470,6 +2473,7 @@ impl com::CommandBuffer<Backend> for CommandBuffer {
         if let Some(sp) = info.subpass {
             let subpass = &sp.main_pass.subpasses[sp.index as usize];
             self.state.target_formats.copy_from(&subpass.target_formats);
+            //self.state.target_samples = subpass.target_samples; //TODO
 
             self.state.target_aspects = Aspects::empty();
             if !subpass.colors.is_empty() {
@@ -3367,6 +3371,7 @@ impl com::CommandBuffer<Backend> for CommandBuffer {
         // we stack the subpasses in the opposite order
         for subpass in render_pass.subpasses.iter().rev() {
             let mut combined_aspects = Aspects::empty();
+            let mut sample_count = 0;
             let descriptor = autoreleasepool(|| {
                 let descriptor = metal::RenderPassDescriptor::new().to_owned();
                 descriptor.set_visibility_result_buffer(Some(&self.shared.visibility.buffer));
@@ -3380,6 +3385,7 @@ impl com::CommandBuffer<Backend> for CommandBuffer {
                     let desc = descriptor.color_attachments().object_at(i as _).unwrap();
 
                     combined_aspects |= Aspects::COLOR;
+                    sample_count = sample_count.max(rat.samples);
                     desc.set_texture(Some(texture));
 
                     if op_flags.contains(native::SubpassOps::LOAD) {
@@ -3404,6 +3410,8 @@ impl com::CommandBuffer<Backend> for CommandBuffer {
                     let rat = &render_pass.attachments[at_id];
                     let texture = framebuffer.attachments[at_id].as_ref();
                     let aspects = rat.format.unwrap().surface_desc().aspects;
+
+                    sample_count = sample_count.max(rat.samples);
                     combined_aspects |= aspects;
 
                     if aspects.contains(Aspects::DEPTH) {
@@ -3445,6 +3453,7 @@ impl com::CommandBuffer<Backend> for CommandBuffer {
                 descriptor,
                 combined_aspects,
                 formats: subpass.target_formats.clone(),
+                sample_count,
             });
         }
 
@@ -3456,11 +3465,14 @@ impl com::CommandBuffer<Backend> for CommandBuffer {
         let sin = self.state.pending_subpasses.pop().unwrap();
 
         self.state.render_pso_is_compatible = match self.state.render_pso {
-            Some(ref ps) => ps.formats == sin.formats,
+            Some(ref ps) => {
+                ps.formats == sin.formats && self.state.target_samples == sin.sample_count
+            }
             None => false,
         };
         self.state.target_aspects = sin.combined_aspects;
         self.state.target_formats.copy_from(&sin.formats);
+        self.state.target_samples = sin.sample_count;
 
         let ds_store = &self.shared.service_pipes.depth_stencil_states;
         let ds_state;
@@ -3514,8 +3526,9 @@ impl com::CommandBuffer<Backend> for CommandBuffer {
             }
         }
 
-        self.state.render_pso_is_compatible =
-            pipeline.attachment_formats == self.state.target_formats;
+        self.state.render_pso_is_compatible = pipeline.attachment_formats
+            == self.state.target_formats
+            && self.state.target_samples == pipeline.samples;
         let set_pipeline = match self.state.render_pso {
             Some(ref ps) if ps.raw.as_ptr() == pipeline.raw.as_ptr() => false,
             Some(ref mut ps) => {