+++ src/backend/metal/src/native.rs
@@ -263,6 +263,7 @@ pub struct GraphicsPipeline {
     pub(crate) vertex_buffers: VertexBufferVec,
     /// Tracked attachment formats
     pub(crate) attachment_formats: SubpassFormats,
+    pub(crate) samples: image::NumSamples,
 }
 
 unsafe impl Send for GraphicsPipeline {}